Move CupertinoLoadingWidget to core module
(cherry picked from commit 2091224f5e55587af2b4c901299673cd0387d15a)
