+++ core/lib/core.dart
@@ -84,6 +84,7 @@ export 'presentation/views/checkbox/labeled_checkbox.dart';
 export 'presentation/views/container/tmail_container_widget.dart';
 export 'presentation/views/clipper/side_arrow_clipper.dart';
 export 'presentation/views/avatar/gradient_circle_avatar_icon.dart';
+export 'presentation/views/loading/cupertino_loading_widget.dart';
 
 // Resources
 export 'presentation/resources/assets_paths.dart';
